server/requirements.txt
grpcio==1.52.0rc1
grpcio-tools==1.51.1
pyntcloud==0.3.1